requirements.txt
Django
psycopg2
gunicorn
whitenoise
python-decouple
dj-database-url
djangorestframework
pytest
pytest-django
model_mommy